+++ mongo/bounties/bounty_queries.js
@@ -28,8 +28,6 @@ db.bounties.find({ season: 1 }).sort({ "reward.amount": -1 });
 
 db.bounties.find({ season: 1 }).sort({ createdAt: -1 });
 
-// Who created the most bounties?
-
 db.bounties
   .aggregate([
     { $match: { season: 1 } },
@@ -58,25 +56,6 @@ db.bounties
   ])
   .pretty();
 
-// (display claimedBy.discordHandle to eyeball)
-
-db.bounties.find({}, { "claimedBy.discordHandle": 1 }).pretty();
-
-// Who completed/submitted the most bounties?
-
-db.bounties
-  .aggregate([
-    { $match: { season: 1 } },
-    {
-      $group: {
-        _id: { submittedBy: "$submittedBy.discordHandle" },
-        totalAmount: { $sum: 1 },
-      },
-    },
-    { $sort: { totalAmount: -1 } },
-  ])
-  .pretty();
-
 // (display submittedBy.discordHandle to eyeball)
 
 db.bounties.find({}, { "submittedBy.discordHandle": 1 }).pretty();
@@ -85,9 +64,13 @@ db.bounties.find({}, { "submittedBy.discordHandle": 1 }).pretty();
 
 // BOUNTY STATUS QUERIES
 
-// How many Bounties were completed on time? (need to add new data fields to consider timeliness)
+// How many Bounties were completed on time?
+// compare these two queries:
 
-// How many Bounties expired? (need to add new data fields to consider expiration)
+db.bounties.find({ $expr: { $gt: ["submittedAt", "dueAt"] } }).count();
+db.bounties.find({ $expr: { $gt: ["dueAt", "submittedAt"] } }).count();
+
+// How many Bounties expired? (TBD; need to add new data fields to consider expiration)
 
 // How many Bounties were expired? (past tense)
 
@@ -124,3 +107,12 @@ db.bounties.find({ "statusHistory.status": "In-Review" }).count();
 // How many Bounties are (currently) 'In-Review'?
 
 db.bounties.find({ status: "In-Review" }).count();
+
+// GUILD SPECIFIC QUERIES (need to add data fields)
+
+// How many bounties were created by each Guild? (need to add new data field)
+// Which Guild created the most bounties?
+// How many bounties were submitted to each Guild?
+// Which Guild received the most submitted bounties?
+// How much bounty reward (in BANK) were given by each Guild?
+// Which Guild gave the most reward?