+++ packages/react-app/docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0-SNAPSHOT
 
 1. Fix CI/CD for preview and production
+2. Add bounty filters
 
 ## 1.0.0-beta2 (2021-08-26)